% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecosim.R
\name{add_level_ecosim_forcing}
\alias{add_level_ecosim_forcing}
\title{Add an additional level for a forcing function time series}
\usage{
add_level_ecosim_forcing(
  factor_set,
  forcing_name,
  level_name,
  forcing_values,
  factor_value = NA
)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{forcing_name}{Name of the forcing function}

\item{level_name}{Name of the new level}

\item{forcing_values}{Numeric vector with the new values for the forcing function. It must have the same length as the existing effort time series for the fleet.}

\item{factor_value}{A single value of the factor used in sensitivity analysis.}
}
\value{
An updated factor set object.
}
\description{
Add an additional level for a forcing function time series
}
