% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecosim.R
\name{add_level_ecosim_shape}
\alias{add_level_ecosim_shape}
\title{Add an additional level for a (environmental response or mediation) shape}
\usage{
add_level_ecosim_shape(
  factor_set,
  shape_name,
  level_name,
  shape_x,
  shape_y,
  factor_value = NA
)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{shape_name}{Name of the shape}

\item{level_name}{Name of the new level}

\item{shape_x}{Numeric vector with the new x values for the shape. Must have a length of 1200 (an Ecosim convention).}

\item{shape_y}{Numeric vector with the new y values for the shape. Must have a length of 1200 (an Ecosim convention).}

\item{factor_value}{A single value of the factor used in sensitivity analysis.}
}
\value{
Updated factor set object.
}
\description{
Add an additional level for a (environmental response or mediation) shape
}
