% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecospace.R
\name{add_level_ecospace_map}
\alias{add_level_ecospace_map}
\title{Add a level to a map in an Ecospace factor set.}
\usage{
add_level_ecospace_map(
  factor_set,
  map_type,
  map_name,
  level_name,
  map_values,
  factor_value = NA
)
}
\arguments{
\item{factor_set}{The factor set.}

\item{map_type}{One of \code{"env_maps"}, \code{"habitats"}, or \code{mpas}}

\item{map_name}{Name of the map for which the level is added. Must already exist in the factor set.}

\item{level_name}{Name of the new level.}

\item{map_values}{Values of the new map.}

\item{factor_value}{Scalar value of the factor representing the magnitude of difference between levels.}
}
\value{
A factor set including the new level for the map.
}
\description{
Add a level to a map in an Ecospace factor set.
}
