% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_analysis.R
\name{calculate_effects_ee}
\alias{calculate_effects_ee}
\title{Calculate elementary effects}
\usage{
calculate_effects_ee(output_name, factor_set, df_ee)
}
\arguments{
\item{output_name}{Name of the output (a column in \code{df_ee} for which elementary effects will be calculated.)}

\item{factor_set}{The factor set from which the elementary effects design was generated.}

\item{df_ee}{A data frame containing the information from the design (run names, IDs, sub-IDs, and a comment stating which factor was changed) as well as the output value of each run.}
}
\value{
A list containing (1) a matrix with the elemntary effect of each factor in each run, (2) mu\* for each factor, (3) sigma for each factor, and (4) a matrix where each row contains mu\* calculated from all trajectories to that point. The latter matrix serves to check if the number of trajectories was sufficient to produce a stable ranking of factors.
}
\description{
Calculate elementary effects
}
