% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_factorial_analysis.R
\name{calculate_effects_ff}
\alias{calculate_effects_ff}
\title{Calculate main and total effects from a full factorial experiment}
\usage{
calculate_effects_ff(output_name, factor_names, df_ff)
}
\arguments{
\item{output_name}{Name of the output column in \code{df_ff}.}

\item{factor_names}{Names of the factor columns in \code{df_ff}.}

\item{df_ff}{Data frame where rows represent runs in a full factorial experiment and columns represent output.}
}
\value{
Data frame with main and total effects.
}
\description{
Calculate main and total effects from a full factorial experiment
}
