% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_values_add}
\alias{change_values_add}
\title{Gradually change values by adding a number}
\usage{
change_values_add(values, summand, start, end)
}
\arguments{
\item{values}{Numeric vector with the original values.}

\item{summand}{The number to add.}

\item{start}{Index where change starts.}

\item{end}{Index where change ends.}
}
\value{
Numeric vector that contains original values before \code{start}, the original values plus \code{summand} from \code{end} onwards, and linearly changed values in between.
}
\description{
Gradually change values by adding a number
}
\examples{
v1=c(2,2,2,2,2)
ecocx::change_values_add(v1,1,2,4)
}
