% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{change_values_mult}
\alias{change_values_mult}
\title{Gradually change values by multiplying with a factor}
\usage{
change_values_mult(values, multi, start, end)
}
\arguments{
\item{values}{Numeric vector with the original values.}

\item{multi}{Multiplier. For example, 1.2 implies a 20 percent increase and 0.8 a 20 percent decrease.}

\item{start}{Index where change starts}

\item{end}{Index where change ends}
}
\value{
Numeric vector that contains original values before \code{start}, \code{multi} times the original values from \code{end} onwards, and linearly changed values in between.
}
\description{
Gradually change values by multiplying with a factor
}
\examples{
v1=c(1,1,1,1,1)
ecocx::change_values_mult(v1,1.5,2,4)
}
