% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_ewe.R
\name{connect_to_ewe}
\alias{connect_to_ewe}
\title{Connect to and execute a test run with the local EwE run console.}
\usage{
connect_to_ewe(path_to_ewe)
}
\arguments{
\item{path_to_ewe}{The path to the run console executable on the local computer.}
}
\value{
An object containing the path, the output folder, test arguments pointing to an internal example  model provided with the package, and whether a test run created the expected results.
}
\description{
Calls the local executable of the EwE run console with example data and checks if expected outputs are generated. This function must be called before any models can be run.
}
