% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_file.R
\name{create_json_changes_ecosim}
\alias{create_json_changes_ecosim}
\title{Create vector with .json file lines for changes.}
\usage{
create_json_changes_ecosim(x, design, factor_set)
}
\arguments{
\item{x}{Row number of the design}

\item{design}{Data frame describing the experiment, e.g., created with an \code{ecocx::sampler_*} method.}

\item{factor_set}{The factor set for the experiment}
}
\value{
Character vector where each element corresponds to a line "Changes" section of the .json file to be generated.
}
\description{
Currently only works for fishing effort.
}
