% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_file.R
\name{create_json_configuration}
\alias{create_json_configuration}
\title{Create a the configuration element of the run console json vector}
\usage{
create_json_configuration(xmlfile, model = NA)
}
\arguments{
\item{xmlfile}{Path to the .eiixml file describing the model}

\item{model}{The model created with \code{create_model_from_xml}. If NA, the modeel is loaded from the provided XML file.}
}
\value{
Character vector where each element corresponds to a line in the .json file to be generated.
}
\description{
This currently works only for Ecosim models.
}
