% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{execute_run}
\alias{execute_run}
\title{Execute a single run from a provided table}
\usage{
execute_run(i, cx_table, ewe_link)
}
\arguments{
\item{i}{Table row dscribing the run.}

\item{cx_table}{Table describing all runs.}

\item{ewe_link}{Link to the EwE Run Console, established with \code{ecocx::connect_to_ewe}.}
}
\description{
Execute a single run from a provided table
}
