% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_ecospace.R
\name{extract_map_data}
\alias{extract_map_data}
\title{Parse map data string from Ecospace EIIXML}
\usage{
extract_map_data(mapdata, basemap)
}
\arguments{
\item{mapdata}{String with map data from EIICML. May be plain or compressed.}

\item{basemap}{The basemap (0 in water cells/the study area, NA elsewhere).}
}
\value{
Matrix with the map data.
}
\description{
Parse map data string from Ecospace EIIXML
}
