% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecospace.R
\name{get_basemap_matrix}
\alias{get_basemap_matrix}
\title{Obtain the model grid and study region}
\usage{
get_basemap_matrix(m)
}
\arguments{
\item{m}{The model.}
}
\value{
Matrix with the model grid and grid cells inside (value \code{0}) and outside (value \code{NA}) the study region.
}
\description{
Obtain the model grid and study region
}
