% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ecopath_ids}
\alias{get_ecopath_ids}
\title{Get Ecopath group IDs from names}
\usage{
get_ecopath_ids(m, names)
}
\arguments{
\item{m}{The model.}

\item{names}{Group names.}
}
\value{
A numeric (integer) vector with group IDs corresponding to the names.
}
\description{
Get Ecopath group IDs from names
}
