% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{get_ecosim_cx_}
\alias{get_ecosim_cx_}
\alias{get_ecosim_cx_biomass}
\alias{get_ecosim_cx_catch}
\title{Obtain time series of biomasses or catch from all Ecosim runs in a computational experiment.}
\usage{
get_ecosim_cx_biomass(cx_table, model, groups = NA)

get_ecosim_cx_catch(cx_table, model, groups = NA)
}
\arguments{
\item{cx_table}{Table describing all runs, as returned by \code{run_ecosim_experiment}.}

\item{model}{The model that generated the outputs to be read. Must have been created with \code{create_model_from_xml}.}

\item{groups}{Character vector with names of the groups for which biomasses should be extracted. If NA (the default), biomasses for all groups are returned.}
}
\value{
Data frame with run name, time step, and biomasses of the groups.
}
\description{
Obtain time series of biomasses or catch from all Ecosim runs in a computational experiment.
}
