% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecosim.R
\name{get_ecosim_effort_length}
\alias{get_ecosim_effort_length}
\title{Get length of an effort time series}
\usage{
get_ecosim_effort_length(factor_set, fleet_name)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{fleet_name}{Name of the fleet}
}
\value{
Length of the time series (i.e., the number of time steps).
}
\description{
Get length of an effort time series
}
