% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecosim.R
\name{get_ecosim_forcing_length}
\alias{get_ecosim_forcing_length}
\title{Get length of a forcing function (time series)}
\usage{
get_ecosim_forcing_length(factor_set, forcing_name)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{forcing_name}{Name of the forcing function}
}
\value{
Length of the forcing function time series (i.e., the number of time steps).
}
\description{
Get length of a forcing function (time series)
}
