% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_ecospace.R
\name{get_ecospace_basemap}
\alias{get_ecospace_basemap}
\title{#' Load map of Ecospace study area from EIIXML}
\usage{
get_ecospace_basemap(xmldoc, scenario_name)
}
\arguments{
\item{xmldoc}{XML2 document.}

\item{scenario_name}{Name of the scenario.}
}
\value{
Object of class \code{EcospaceMap} containing information about the map (e.g., number of rows and columns) and depth values.
}
\description{
#' Load map of Ecospace study area from EIIXML
}
