% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_ecospace.R
\name{get_ecospace_mpamaps}
\alias{get_ecospace_mpamaps}
\title{Load static MPA maps from EIIXML}
\usage{
get_ecospace_mpamaps(xmldoc, scenario_name)
}
\arguments{
\item{xmldoc}{XML2 document.}

\item{scenario_name}{Name of the scenario.}
}
\value{
List of \code{EcospaceMap} objects, each containing one MPA.
}
\description{
Load static MPA maps from EIIXML
}
