% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xml_ecospace.R
\name{get_ecospace_scenarios}
\alias{get_ecospace_scenarios}
\title{Load Ecospace scenario information from EIIXML}
\usage{
get_ecospace_scenarios(xmldoc)
}
\arguments{
\item{xmldoc}{XML2 document.}
}
\value{
Data frame with Ecospace scenario names and IDs.
}
\description{
Load Ecospace scenario information from EIIXML
}
