% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_factor_by_name}
\alias{get_factor_by_name}
\title{Get a factor from a factor set by name}
\usage{
get_factor_by_name(factor_set, name)
}
\arguments{
\item{factor_set}{A factor set.}

\item{name}{Name of the factor.}
}
\value{
Object representing the factor.
}
\description{
Get a factor from a factor set by name
}
