% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_path_to_exampledata}
\alias{get_path_to_exampledata}
\title{Obtain the path of the folder with raw example data}
\usage{
get_path_to_exampledata()
}
\value{
A string with the path.
}
\description{
Obtain the path of the folder with raw example data
}
