% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_run_console_link}
\alias{get_run_console_link}
\title{Obtain a link for downloading the EwE Run Console}
\usage{
get_run_console_link()
}
\value{
Link to a release of the Run Console that the package was tested with.
}
\description{
Each version of EcoCX is tested against a specific release of the EwE Run Console. This function returns a link to that release.
}
