% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_seq_ids}
\alias{get_seq_ids}
\title{Get Ecopath sequential IDs from names}
\usage{
get_seq_ids(m, names)
}
\arguments{
\item{m}{The model.}

\item{names}{Group names.}
}
\value{
A numeric (integer) vector with sequential IDs corresponding to the names.
}
\description{
Get Ecopath sequential IDs from names
}
