% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lut_tscode}
\alias{lut_tscode}
\title{Obtain information about types of time series}
\usage{
lut_tscode(code = NULL)
}
\arguments{
\item{code}{The numerical time series code or NULL.}
}
\value{
If a code was provided, a data frame with one row containing information about the code provided. If the code was NULL, a look-up table with codes and associated information.
}
\description{
'lut_tscode' provides information about time series types, like descriptive names and units, from an Ecosim-specific code, for example, whether they are drivers (forcing certain values at model runtime) or reference time series (for model fitting).
}
\examples{
lut_tscode(1)
lut_tscode("3")
lut_tscode()
}
