% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecosim.R
\name{new_ecosim_factor_set}
\alias{new_ecosim_factor_set}
\title{Create a set of factors that vary between Ecosim model runs}
\usage{
new_ecosim_factor_set(m, default_name = "default")
}
\arguments{
\item{m}{An Ecosim model object created with \code{load_model_from_xml()}.}

\item{default_name}{Name that the default value}
}
\value{
A list of factors like fishing effort and environmental time series with their default values in the model.
}
\description{
The factor set contain lists of alternative fishing effort, environmental response shapes, and other potential factors that might vary between model runs.
First, use this function to generate a factor set with one level per factor,then add alternatives with add_ecosim_factor_level().
}
\examples{
xmlfile=paste0(ecocx::get_path_to_exampledata(),'anchovy_bay_ecosim_ex.eiixml')
m <- ecocx::load_model_from_xml(xmfile)
flist <- new_ecosim_factor_set(m)
}
