% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecospace.R
\name{new_ecospace_factor_set}
\alias{new_ecospace_factor_set}
\title{Create a set of factors that vary between Ecosim model runs}
\usage{
new_ecospace_factor_set(m, default_name = "default")
}
\arguments{
\item{m}{A model created with \code{load_model_from_xml()}.}

\item{default_name}{Name that the default value}
}
\value{
A list of Ecosim and Ecospace factors like fishing effort time series and MPA maps with their default values in the model. Only Ecosim factors that transfer to Ecospace simulations are included.
}
\description{
The factor set contain lists of alternative fishing effort, environmental response shapes, and other potential factors that might vary between model runs.
First, use this function to generate a factor set with one level per factor,then add alternatives with add_ecosim_factor_level().
}
\examples{
xmlfile=paste0(ecocx::get_path_to_exampledata(),'anchovy_bay_ecospace_ex.eiixml')
m <- ecocx::load_model_from_xml(xmfile,ecospace_scenario="BayOfAnchovies")
fset <- new_ecospace_factor_set(m)
}
