% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{plot.ecocx_factor_set}
\alias{plot.ecocx_factor_set}
\title{Plot function for factor sets}
\usage{
\method{plot}{ecocx_factor_set}(x, ...)
}
\arguments{
\item{...}{Unused

Shows one plot for each option in the factor set.}

\item{object}{The factor set}
}
\value{
Nothing.
}
\description{
Plot function for factor sets
}
