% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization.R
\name{plot_all_runs}
\alias{plot_all_runs}
\title{Plot all runs from a computational experiment}
\usage{
plot_all_runs(df_cx, run_names = NA, timesteps = NA, outputs = NA)
}
\arguments{
\item{df_cx}{Dataframe with experiment outputs. Must have at least 3 columns: \code{run_name}, \code{timestep}, and at least one column with model outputs (e.g., one column per group biomass).}

\item{run_names}{Optional vector with names of the runs to plot. Must match run names in \code{df_cx}.}

\item{timesteps}{Optional vector with timesteps to plot. Best to not skip timesteps; e.g., use \code{timesteps=1:100} to plot the first 100 timesteps only.}

\item{outputs}{Optional vector with outputs like group or fleet names to plot. Must match output names (column names) in \code{df_cx}.}
}
\description{
Plots one line for each run and output (e.g., biomasses), color-coded by output. Inflexible and slow but fine for quick plotting of a few (<100) runs. For faster and more complex plotting, better use custom packages like \code{ggplot2}.
}
