% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization.R
\name{plot_ecosim_factor_levels}
\alias{plot_ecosim_factor_levels}
\title{Plot all levels of a factor}
\usage{
plot_ecosim_factor_levels(factor_set, name)
}
\arguments{
\item{factor_set}{The factor set.}

\item{name}{Name of the factor: Must refer to a forcing function/environmental driver time series, mediation or environmental response shape, or fishing effort time series.}
}
\value{
A base R plot.
}
\description{
Plot all levels of a factor
}
