% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization.R
\name{plot_ecospace_map}
\alias{plot_ecospace_map}
\title{Plot an Ecospace map}
\usage{
plot_ecospace_map(map)
}
\arguments{
\item{map}{EcospaceMap object.}
}
\value{
Nothing.
}
\description{
Plot an Ecospace map
}
