% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization.R
\name{plot_ee}
\alias{plot_ee}
\title{Plot results of the elementary effects method}
\usage{
plot_ee(ee_results)
}
\arguments{
\item{ee_results}{Object returned by \code{calculate_effects_ee}.}
}
\value{
Nothing.
}
\description{
Plot results of the elementary effects method
}
