% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization.R
\name{plot_ee_mustar}
\alias{plot_ee_mustar}
\title{Plot mu_star as a function of trajectories completed}
\usage{
plot_ee_mustar(ee_results)
}
\arguments{
\item{ee_results}{Object returned by \code{calculate_effects_ee}.}
}
\value{
Nothing.
}
\description{
Use this to check if you evaluated enough trajectories to get a stable factor ranking.
}
