% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization.R
\name{plot_ff_effects}
\alias{plot_ff_effects}
\title{Plot direct and interaction effect sizes from a full factorial experiment.}
\usage{
plot_ff_effects(ff_results, all = FALSE)
}
\arguments{
\item{ff_results}{Results from a full factorial experiment, calculated with \code{calculate_effects_ff}.}

\item{all}{Boolean. Shall all factors be plotted, even those with only one level?}
}
\value{
Nothing.
}
\description{
Plot direct and interaction effect sizes from a full factorial experiment.
}
