% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecosim.R
\name{remove_level_ecosim_effort}
\alias{remove_level_ecosim_effort}
\title{Remove a level for the effort time series of a fleet.}
\usage{
remove_level_ecosim_effort(factor_set, fleet_name, level_name)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{fleet_name}{Name of the fleet}

\item{level_name}{Name of the level to remove}
}
\value{
An updated factor set object.
}
\description{
At least one level must remain at all times. To replace the last level, add a new level first, then remove the old one.
}
