% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecosim.R
\name{remove_level_ecosim_forcing}
\alias{remove_level_ecosim_forcing}
\title{Remove a level for a forcing time series.}
\usage{
remove_level_ecosim_forcing(factor_set, forcing_name, level_name)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{forcing_name}{Name of the forcing function}

\item{level_name}{Name of the level to remove}
}
\value{
An updated factor set object.
}
\description{
At least one level must remain at all times. To replace the last level, add a new level first, then remove the old one.
}
