% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors_ecosim.R
\name{remove_level_ecosim_vulnerability}
\alias{remove_level_ecosim_vulnerability}
\title{Remove an level for the vulnerability table.}
\usage{
remove_level_ecosim_vulnerability(factor_set, level_name)
}
\arguments{
\item{factor_set}{A factor set, e.g.,created with new_ecosim_factor_set().}

\item{level_name}{Name of the level to remove.}
}
\value{
An updated factor set object.
}
\description{
At least one level must remain at all times. To replace the last level, add a new level first, then remove the old one.
}
