% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{remove_option_ecosim_effort}
\alias{remove_option_ecosim_effort}
\title{Remove an option for the effort time series of a fleet.}
\usage{
remove_option_ecosim_effort(factor_set, fleet_name, option_name)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{fleet_name}{Name of the fleet}

\item{option_name}{Name of the option to remove}
}
\value{
An updated factor set object.
}
\description{
At least one option must remain at all times. To replace the last option, add a new option first, then remove the olds one.
}
