% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{remove_option_ecosim_forcing}
\alias{remove_option_ecosim_forcing}
\title{Remove an option for a forcing time series.}
\usage{
remove_option_ecosim_forcing(factor_set, forcing_name, option_name)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{forcing_name}{Name of the forcing function}

\item{option_name}{Name of the option to remove}
}
\value{
An updated factor set object.
}
\description{
At least one option must remain at all times. To replace the last option, add a new option first, then remove the old one.
}
