% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{remove_option_ecosim_vulnerability}
\alias{remove_option_ecosim_vulnerability}
\title{Remove an option for the vulnerability table.}
\usage{
remove_option_ecosim_vulnerability(factor_set, option_name)
}
\arguments{
\item{factor_set}{A factor set, e.g.,created with new_ecosim_factor_set().}

\item{option_name}{Name of the option to remove.}
}
\value{
An updated factor set object.
}
\description{
At least one option must remain at all times. To replace the last option, add a new option first, then remove the old one.
}
