% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_ecosim_experiment}
\alias{run_ecosim_experiment}
\title{Run a computational experiment}
\usage{
run_ecosim_experiment(
  design,
  xml_model_path,
  factor_set,
  model,
  ewe_link,
  out_folder,
  parallel = F
)
}
\arguments{
\item{design}{Data frame describing the experiment, e.g., created with an \code{ecocx::sampler_*} method.}

\item{xml_model_path}{Path to XML file describing the model.}

\item{factor_set}{The factor set for the experiment}

\item{ewe_link}{Link to the EwE Run Console generated with \code{ecocx::connect_to_ewe}}

\item{out_folder}{Outputs from each run will be stored in subfolders of this folder.}

\item{parallel}{Whether to use in case of parallel processing. If \code{true}, you need the \code{future} and \code{future.apply} package installed and call future::plan yourself.}

\item{output_fun}{A named list of functions to calculate output indicators. Names will be used in the generated output table.}
}
\value{
A data frame with information about the runs, e.g., where outputs are stored.
}
\description{
Takes a pre-generated description of the required runs and exceutes them. Stores results on disk in one folder per run, which also contains a generated Run Console .json file and the Run Console log. Finally, calculates outputs using the provided functions,
which can be default functions from this package or user-generated.
}
