% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sample_option_from_list}
\alias{sample_option_from_list}
\title{Sample one entry from a list, returning the index}
\usage{
sample_option_from_list(factor)
}
\arguments{
\item{factor}{An ecocx factor (not necessarily a factor in the R sense) from with one or more options.}
}
\value{
Index of the option.
}
\description{
Sample one entry from a list, returning the index
}
