% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization.R
\name{tol_colors}
\alias{tol_colors}
\title{Obtain a vector with hex color codes for Paul Tol's palettes.}
\usage{
tol_colors(pal = "muted", black = F)
}
\arguments{
\item{pal}{Palette to return. One of "bright", "vibrant", "high" (contrast), "muted".}

\item{black}{Should black be added as an additional color?}
}
\value{
A vector of color hex codes.
"export
}
\description{
Obtain a vector with hex color codes for Paul Tol's palettes.
}
