% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{add_level_ecosim_effort}
\alias{add_level_ecosim_effort}
\title{Add an additional level for the fishing effort time series of a fleet}
\usage{
add_level_ecosim_effort(
  factor_set,
  fleet_name,
  level_name,
  effort_values,
  factor_value = NA
)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{fleet_name}{Name of the fleet}

\item{level_name}{Name for the new level}

\item{effort_values}{A numeric vector with the new effort values. It must have the same length as the existing effort time series for the fleet.}

\item{factor_value}{A single value of the factor used in sensitivity analysis. For example, if this doubles fishing effort compared to a baseline effort, this parameter should be 2, and for the baseline level, 1. Used in elementary effects method and Sobol indices.}
}
\value{
An updated factor set object.
}
\description{
Add an additional level for the fishing effort time series of a fleet
}
