% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{add_level_ecosim_vulnerability}
\alias{add_level_ecosim_vulnerability}
\title{Add an alternative vulnerability matrix.}
\usage{
add_level_ecosim_vulnerability(factor_set, level_name, v_matrix)
}
\arguments{
\item{factor_set}{A factor set, e.g., created with new_ecosim_factor_set().}

\item{level_name}{Name of the new level.}

\item{v_matrix}{New vulnerability matrix. It must have the same dimensions as the original model's vulnerability matrix, contain numbers where the original model's matrix contains numbers, and NAs where the original model's matrix contains NAs.}
}
\value{
An updated factor set including the new level.
}
\description{
Add an alternative vulnerability matrix.
}
