% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{add_option_ecosim_effort}
\alias{add_option_ecosim_effort}
\title{Add an additional option for the fishing effort time series of a fleet}
\usage{
add_option_ecosim_effort(factor_set, fleet_name, option_name, effort_values)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{fleet_name}{Name of the fleet}

\item{option_name}{Name for the new option}

\item{effort_values}{A numeric vector with the new effort values. It must have the same length as the existing effort time series for the fleet.}
}
\value{
An updated factor set object.
}
\description{
Add an additional option for the fishing effort time series of a fleet
}
