% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{add_option_ecosim_forcing}
\alias{add_option_ecosim_forcing}
\title{Add an additional option for a forcing function time series}
\usage{
add_option_ecosim_forcing(
  factor_set,
  forcing_name,
  option_name,
  forcing_values,
  factor_value
)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{forcing_name}{Name of the forcing function}

\item{option_name}{Name of the new option}

\item{forcing_values}{Numeric vector with the new values for the forcing function. It must have the same length as the existing effort time series for the fleet.}

\item{factor_value}{A single value of the factor used in sensitivity analysis. For example, if this options represents a 2 degree temperetaure change over a baseline option, this parameter should be 2, vs. 0 for the baseline.}
}
\value{
An updated factor set object.
}
\description{
Add an additional option for a forcing function time series
}
