% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{add_option_ecosim_shape}
\alias{add_option_ecosim_shape}
\title{Add an additional option for a (environmental response or mediation) shape}
\usage{
add_option_ecosim_shape(factor_set, shape_name, option_name, shape_x, shape_y)
}
\arguments{
\item{factor_set}{A factor set created with new_ecosim_factor_set()}

\item{shape_name}{Name of the shape}

\item{option_name}{Name of the new option}

\item{shape_x}{Numeric vector with the new x values for the shape. Must have a length of 1200 (an Ecosim convention).}

\item{shape_y}{Numeric vector with the new y values for the shape. Must have a length of 1200 (an Ecosim convention).}
}
\value{
Updated factor set object.
}
\description{
Add an additional option for a (environmental response or mediation) shape
}
