% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_file.R
\name{create_json_ecosim_run}
\alias{create_json_ecosim_run}
\title{Create the "EcosimRun" element of the run console .json file}
\usage{
create_json_ecosim_run()
}
\value{
Character vector where each element corresponds to a line in the .json file to be generated.
}
\description{
Create the "EcosimRun" element of the run console .json file
}
