% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_file.R
\name{create_json_vector}
\alias{create_json_vector}
\title{Create a json vector from a row in a cx_table}
\usage{
create_json_vector(x, cx_table, design, factor_set)
}
\arguments{
\item{x}{Row number in the provided \code{cx_table}.}

\item{cx_table}{A data frame  with columns 'run_name' (matching the names in the \code{design}), 'model' (path to the EIIXML file), 'folder' (base folder of the run), 'json' (path of the .json file to be created)}

\item{design}{Data frame describing the experiment, e.g., created with an \code{ecocx::sampler_*} method.}

\item{factor_set}{The factor set for the experiment.}
}
\value{
A character vector where each element is a line of the .json file.
}
\description{
Create a json vector from a row in a cx_table
}
