% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{get_factor_scalar_values}
\alias{get_factor_scalar_values}
\title{Get scalar factor values from a factor set}
\usage{
get_factor_scalar_values(factor_set)
}
\arguments{
\item{factor_set}{A factor set.}
}
\value{
Data frame containing all factors and levels with their values.
}
\description{
Some sensitivity analysis methods (e.g., elementary effects, Sobol) require or benefit from (e.g., machine learning using ordered instead of categorical factor levels) each factor to have a scalar value representing the magnitude of change between levels.
}
