% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_results.R
\name{get_single_output}
\alias{get_single_output}
\alias{get_ecosim_run_biomass}
\alias{get_ecosim_run_catch}
\title{Obtain time series of biomasses or catch from a single Ecosim run}
\usage{
get_ecosim_run_biomass(folder, model, groups = NA)

get_ecosim_run_catch(folder, model, groups = NA)
}
\arguments{
\item{folder}{Path to folder containing the model outputs. If the outputs are not found, the function searches subfolders and uses the first fitting output file found.}

\item{model}{The model that generated the outputs to be read. Must have been created with \code{create_model_from_xml}.}

\item{groups}{Character vector with names of the groups for which biomasses should be extracted. If NA (the default), biomasses for all groups are returned.}
}
\value{
Data frame with a timestep column and one column for each group listed in \code{biomass}.
}
\description{
Obtain time series of biomasses or catch from a single Ecosim run
}
