% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model_from_xml.R
\name{load_model_from_xml}
\alias{load_model_from_xml}
\title{Load an EwE model from an XML file exported with the Ecopath software}
\usage{
load_model_from_xml(xmlfile)
}
\arguments{
\item{xmlfile}{Path to .xml file.}
}
\value{
An object of class 'ecocx_model'. Its elements contain core information about functional groups and their parameters, fisheries, mediation and environmental response functions, and other information. Currently, only
Ecopath and Ecosim (but not Ecospace) information is loaded. Furthermore, the package currently supports only models with one Ecosim scenario.
}
\description{
Load an EwE model from an XML file exported with the Ecopath software
}
\examples{
m <- ecocx::load_model_from_xml(xmlfile=paste0(ecocx::get_path_to_exampledata(),'anchovy_bay_ecosim_ex.eiixml'))
m$ecopath$basic_estimates
m$ecosim$shapes$Tempcold$x[1:5]
m$ecosim$shapes$Tempcold$y[1:5]
}
