% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizualization.R
\name{plot_ecosim_output_variance}
\alias{plot_ecosim_output_variance}
\title{Plot the variance of Ecosim outputs}
\usage{
plot_ecosim_output_variance(df_cx, timestep = NA, outputs = NA)
}
\arguments{
\item{df_cx}{Data frame with the outputs}
}
\description{
Plot the variance of Ecosim outputs
}
