% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampler_full_factorial}
\alias{sampler_full_factorial}
\title{Create a full factorial (cpmoutational) experiment}
\usage{
sampler_full_factorial(factor_set)
}
\arguments{
\item{factor_set}{A factor set including all factor levels to combine.}
}
\value{
A data frame with 'size' rows and automatically generated run ID, sub-ID, run_name and comment columns; as well as one column per factor. Each row contains one combination of factors. Together, the data frame covers all possible combinations
}
\description{
Creates a data frame with all factor combinations in a factor set.
}
