% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplers.R
\name{sampler_random}
\alias{sampler_random}
\title{Randomly sample factor levels for standard Monte Carlo runs}
\usage{
sampler_random(factor_set, size = 1)
}
\arguments{
\item{factor_set}{A factor set including all options to sample from.}

\item{size}{Sample size (the number of Monte Carlo runs)}
}
\value{
A data frame with automatically generated run ID, sub-ID, run_name and comment columns; as well as one column per factor with a random sample.
}
\description{
Randomly sample factor levels for standard Monte Carlo runs
}
