% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{set_factor_scalar_values}
\alias{set_factor_scalar_values}
\title{Set scalar factor values in a factor set}
\usage{
set_factor_scalar_values(factor_set, values_table)
}
\arguments{
\item{factor_set}{A factor set.}

\item{values_table}{A data frame created with \code{get_factor_scalar_values}. The column factor_value should be set to the new values.}
}
\value{
Factor set with the set values.
}
\description{
Some sensitivity analysis methods (e.g., elementary effects, Sobol) require or benefit from (e.g., machine learning using ordered instead of categorical factor levels) each factor to have a scalar value representing the magnitude of change between levels.
}
