% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_factors.R
\name{summary.ecocx_factor_set}
\alias{summary.ecocx_factor_set}
\title{Summary function for factor sets}
\usage{
\method{summary}{ecocx_factor_set}(object, ...)
}
\arguments{
\item{object}{The factor set}

\item{...}{Unused}
}
\value{
A data frame with the number of levels per factor.
}
\description{
Summary function for factor sets
}
