% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_json_file.R
\name{write_json_file}
\alias{write_json_file}
\title{Create and write the Run Console json file from a vector}
\usage{
write_json_file(i, cx_table, design, factor_set)
}
\arguments{
\item{i}{Row of the \code{cx_table} for which to create the file}

\item{cx_table}{A data frame  with columns 'run_name' (matching the names in the \code{design}), 'model' (path to the EIIXML file), 'folder' (base folder of the run), 'json' (path of the .json file to be created)}

\item{design}{Data frame describing the experiment, e.g., created with an \code{ecocx::sampler_*} method.}

\item{factor_set}{The factor set for the experiment.}
}
\value{
A list of NULLs. Ignore the output.
}
\description{
Create and write the Run Console json file from a vector
}
