% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_vulnerability_csv}
\alias{write_vulnerability_csv}
\title{Write a vulnerability table to disk for loading in the .json file.}
\usage{
write_vulnerability_csv(table, path)
}
\arguments{
\item{table}{The vulneability table.}

\item{path}{Where should it be saved?}
}
\value{
Nothing.
}
\description{
Write a vulnerability table to disk for loading in the .json file.
}
